ShuttleAutomation moved to lib-command-parameters
