+++ src/commands/AudacityCommand.cpp
@@ -34,7 +34,7 @@ ShuttleGui.
 
 #include "ConfigInterface.h"
 
-#include "../ShuttleAutomation.h"
+#include "ShuttleAutomation.h"
 #include "../ShuttleGui.h"
 #include "../widgets/ProgressDialog.h"
 #include "../widgets/HelpSystem.h"
