+++ src/effects/AutoDuck.h
@@ -12,7 +12,7 @@
 #define __AUDACITY_EFFECT_AUTODUCK__
 
 #include "Effect.h"
-#include "../ShuttleAutomation.h"
+#include "ShuttleAutomation.h"
 #include <float.h> // for DBL_MAX
 #include "../widgets/wxPanelWrapper.h"
