+++ src/effects/ChangeSpeed.h
@@ -14,7 +14,7 @@
 #define __AUDACITY_EFFECT_CHANGESPEED__
 
 #include "Effect.h"
-#include "../ShuttleAutomation.h"
+#include "ShuttleAutomation.h"
 
 class wxSlider;
 class wxChoice;
