+++ src/effects/ChangeTempo.h
@@ -22,7 +22,7 @@
 #endif
 
 #include "SoundTouchEffect.h"
-#include "../ShuttleAutomation.h"
+#include "ShuttleAutomation.h"
 
 class wxSlider;
 class wxCheckBox;
