+++ src/effects/Compressor.h
@@ -12,7 +12,7 @@
 #define __AUDACITY_EFFECT_COMPRESSOR__
 
 #include "TwoPassSimpleMono.h"
-#include "../ShuttleAutomation.h"
+#include "ShuttleAutomation.h"
 #include "MemoryX.h"
 #include "../widgets/wxPanelWrapper.h"
