+++ src/effects/Effect.cpp
@@ -27,7 +27,7 @@
 #include "../LabelTrack.h"
 #include "../ProjectSettings.h"
 #include "../SelectFile.h"
-#include "../ShuttleAutomation.h"
+#include "ShuttleAutomation.h"
 #include "../ShuttleGui.h"
 #include "../SyncLock.h"
 #include "ViewInfo.h"
