+++ src/effects/Paulstretch.h
@@ -11,7 +11,7 @@
 #define __AUDACITY_EFFECT_PAULSTRETCH__
 
 #include "Effect.h"
-#include "../ShuttleAutomation.h"
+#include "ShuttleAutomation.h"
 #include <float.h> // for FLT_MAX
 
 class ShuttleGui;
