+++ src/effects/ScienFilter.h
@@ -18,7 +18,7 @@ Vaughan Johnson (Preview)
 #include "Biquad.h"
 
 #include "StatefulPerTrackEffect.h"
-#include "../ShuttleAutomation.h"
+#include "ShuttleAutomation.h"
 #include "../widgets/wxPanelWrapper.h"
 #include <float.h> // for FLT_MAX
