+++ src/effects/StatefulEffectBase.cpp
@@ -24,8 +24,8 @@
 #include "../LabelTrack.h"
 #include "../ProjectSettings.h"
 #include "../SelectFile.h"
-#include "../ShuttleAutomation.h"
-#include "../ShuttleGui.h"
+#include "ShuttleAutomation.h"
+#include "ShuttleGui.h"
 #include "../SyncLock.h"
 #include "ViewInfo.h"
 #include "WaveTrack.h"
