+++ src/effects/nyquist/Nyquist.cpp
@@ -61,7 +61,7 @@ effects from this one class.
 #include "PluginManager.h"
 #include "Project.h"
 #include "ProjectRate.h"
-#include "../../ShuttleAutomation.h"
+#include "ShuttleAutomation.h"
 #include "../../ShuttleGetDefinition.h"
 #include "../../ShuttleGui.h"
 #include "TempDirectory.h"
